App-wide default keep mode via Appyx singleton
